% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossnma.heatplot.R
\name{crossnma.heatplot}
\alias{crossnma.heatplot}
\title{Heat Plot}
\usage{
crossnma.heatplot(
  x,
  central.tdcy = "median",
  exp = TRUE,
  order = NULL,
  low.colour = "red",
  mid.colour = "white",
  high.colour = "springgreen4",
  prt.cov1.value = NULL,
  prt.cov2.value = NULL,
  prt.cov3.value = NULL,
  digits = 2,
  cell.text.size = 12,
  trt.name.size = 20,
  axis.title.size = 12
)
}
\arguments{
\item{x}{A \code{crossrun} object produced by running \code{crossnma.run()}.}

\item{central.tdcy}{The statistic to be used to measure the relative treatment effects. The options are "mean" and "median".}

\item{exp}{If TRUE (default), odds ratios are displayed. If FALSE, log odds ratios will be presented.}

\item{order}{A vector of treatment names (character) representing the order in which to display these treatments.}

\item{low.colour}{A string indicating the colour of low relative treatment effects for the heat plot (e.g odds ratio of ~0.5)}

\item{mid.colour}{A string indicating the colour of null relative treatment effects for the heat plot (e.g odds ratio of ~1.0).}

\item{high.colour}{A string indicating the colour of high relative treatment effects for the heat plot (e.g odds ratio of ~2.0).}

\item{prt.cov1.value}{The participant covariate value of \code{cov1} for which to report the results. Must be specified for network meta-regression and when individual participant dataset is used in the analysis.
For dichotomous covariates, a character of the level (used in the data) should be indicated.}

\item{prt.cov2.value}{The participant covariate value of \code{cov2} for which to report the results. Must be specified for network meta-regression and when individual participant dataset is used in the analysis.
For dichotomous covariates, a character of the level (used in the data) should be indicated.}

\item{prt.cov3.value}{The participant covariate value of \code{cov3} for which to report the results. Must be specified for network meta-regression and when individual participant dataset is used in the analysis.
For dichotomous covariates, a character of the level (used in the data) should be indicated.}

\item{digits}{The number of digits to be used when displaying the results}

\item{cell.text.size}{The size of the cell entries; the relative treatment effect and the 95\% credible intervals}

\item{trt.name.size}{The size of treatment names placed on the top and left of the plot}

\item{axis.title.size}{The size of titles placed on the top and left of the plot}
}
\value{
\code{heatplot}  League heat plot, where a color scale is used to represent the relative treatment effects.
}
\description{
Produces a heat plot that contain point estimates of relative effects
for all possible pairs of treatments along with 95\% credible intervals obtained with the quantile method.
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

#=========================#
# Create a jags model     #
#=========================#
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,
  reference = "A", trt.effect = "random", method.bias = "naive")

#=========================#
# Fit jags model          #
#=========================#
fit <-
  crossnma.run(model = mod, n.adapt = 20,
    n.iter = 50, thin = 1, n.chains = 3)

#=========================#
# Create the league table #
#=========================#
crossnma.heatplot(fit)

}
\seealso{
\code{\link{crossnma.run}}
}
