% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossnma.R
\name{crossnma}
\alias{crossnma}
\title{Run JAGS to fit cross NMA and NMR}
\usage{
crossnma(
  x,
  inits = NULL,
  n.burnin = floor(n.iter/2),
  n.iter = 2000,
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  n.chains = 2,
  monitor = NULL
)
}
\arguments{
\item{x}{An object produced by \code{\link{crossnma.model}}.}

\item{inits}{A list with n.chains elements; each element is itself a list of initial values for each model parameter,
or a function that generates starting values. Default is NULL where jags.parallel() function
will internally generate starting values for each parameter.}

\item{n.burnin}{Number of burnin iterations for the MCMC chains. Default is n.iter/2 which discards the first half of the iterations.}

\item{n.iter}{Number of iterations to run each MCMC chain. Default is 2000.}

\item{n.thin}{Number of thinning for the MCMC chains. Default is max(1, floor((n.iter - n.burnin) / 1000)),
that is only thinning if there are more than 2000 iterations.}

\item{n.chains}{Number of MCMC chains. Default is 2.}

\item{monitor}{A character vector of the names of the parameters to
be monitored. Basic parameters (depends on the analysis) will be automatically monitored
and only additional parameters need to be specified.
Default is NULL.}
}
\value{
An object of class \code{crossnma} which is a list containing the
following components:
\item{jagsfit}{ An rjags-class object produced when R2jags package used to run the JAGS
  model.}
\item{model}{The \code{crossnma.model} object obtained from
  \code{\link{crossnma.model}} which was used to run JAGS.}
\item{trt.key}{A table of treatment names and their correspondence
  to integers used in the JAGS model.}
\item{call}{Function call.}
\item{version}{Version of R package \bold{crossnma} used to create
  object.}
}
\description{
This function takes the JAGS model from an object produced by
\code{\link{crossnma.model}} and runs it using \code{jags.parallel} in
R2jags package.
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

# Create a JAGS model
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,sm="OR",
  reference = "A", trt.effect = "random", method.bias = "naive")

# Fit JAGS model
fit <-crossnma(mod,
n.burnin =10,n.iter = 50, n.thin = 1, n.chains = 3)

# Display the output
summary(fit)
plot(fit)

}
\seealso{
\code{\link{crossnma.model}},
  \code{\link[R2jags]{jags.parallel}}
}
\author{
Tasnim Hamza \email{tasnim.hamza@ispm.unibe.ch}, Guido
  Schwarzer \email{sc@imbi.uni-freiburg.de}
}
