% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossnma.run.R
\name{crossnma.run}
\alias{crossnma.run}
\title{Run JAGS to fit cross NMA and NMR}
\usage{
crossnma.run(
  model,
  n.adapt = 1000,
  n.burnin = floor(n.iter/2),
  n.iter,
  thin = 1,
  n.chains = 2,
  quiet = TRUE,
  monitor = NULL
)
}
\arguments{
\item{model}{A \code{crossnmaModel} object produced by
\code{\link{crossnma.model}}.}

\item{n.adapt}{Number of adaptations for the MCMC chains. Default
is 1000.}

\item{n.burnin}{Number of burnin iterations for the MCMC chains.}

\item{n.iter}{Number of iterations for the MCMC chains.}

\item{thin}{Number of thinning for the MCMC chains. Default is 1.}

\item{n.chains}{Number of MCMC chains. Default is 2.}

\item{quiet}{A logical. If TRUE (default), the warning message will
not be displayed.}

\item{monitor}{A vector of additional parameters to
monitor. Default is NULL.}
}
\value{
\code{crossnma.run} returns an object of class
  \code{crossrun} which is a list containing the following
  components:

\code{samples} The MCMC samples produced by running the
  JAGS model.

\code{model} The \code{crossnmaModel} object obtained from
  \code{\link{crossnma.model}} which was used to run \code{jags}.

\code{trt.key} A table of the treatment names and their
  correspondence to integers used in the JAGS model.
}
\description{
This function takes the JAGS model from an object
  produced by \code{\link{crossnma.model}} and runs it using
  \code{jags.model} in rjags package.
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

#=========================#
# Create a jags model     #
#=========================#
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,
  reference = "A", trt.effect = "random", method.bias = "naive")

#=========================#
# Fit jags model          #
#=========================#
fit <-
  crossnma.run(model = mod, n.adapt = 20,
    n.iter = 50, thin = 1, n.chains = 3)

#=========================#
# Display the output      #
#=========================#
summary(fit)
plot(fit)


}
\seealso{
\code{\link{crossnma.model}}, \code{\link{jags.model}}
}
