% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ipddata}
\alias{ipddata}
\title{Simulated individual participant dataset.}
\format{
A data frame with 2950 rows and 10 variables:
\describe{
  \item{id}{numeric, study identifier}
  \item{relapse}{binary data, respond indicator, 0=no relapse and 1=relapse}
  \item{treat}{character, indicating the assigned treatment to each participant}
  \item{design}{character, design of the study, either 'rct' or 'nrs'}
  \item{age}{numeric, age of the participant}
  \item{sex}{binary data, sex of the participant, 0=Female and 1=Male}
  \item{rob}{character, the risk of bias of the study, 'low', 'high','unclear'}
  \item{unfavored}{numeric, the indicator of the unfavored treatment in each study, values are 0 or 1}
  \item{bias.group}{numeric, the bias effect of the study, 1 = if the study has inactive treatment and adjust for bias effect, 2= if the study has active treatments and it is assumed another bias effect, 0=no bias adjustment}
  \item{year}{numeric, the year study published}

}
}
\usage{
ipddata
}
\description{
A dataset containing 2950 participants who are treated in four
different treatments: A, B, C and D.  The dataset includes four
studies. The outcome is binary. There are 10 attributes on
individual level.
}
\keyword{datasets}
