% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgraph.crossnmaModel.R
\name{netgraph.crossnmaModel}
\alias{netgraph.crossnmaModel}
\title{Produce a network plot}
\usage{
\method{netgraph}{crossnmaModel}(x, ...)
}
\arguments{
\item{x}{A \code{crossnmaModel} object produced by \code{crossnma.model}.}

\item{...}{\dots Additional arguments from \code{\link{netgraph.netmeta}}}
}
\description{
This function creates a plot of evidence network using an object produced by \code{crossnma.model}.
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

#=========================#
# Create a jags model     #
#=========================#
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,
  reference = "A", trt.effect = "random", method.bias = "naive")

#=========================#
# Fit jags model          #
#=========================#
fit <-
  crossnma.run(model = mod, n.adapt = 20,
    n.iter = 50, thin = 1, n.chains = 3)

#=========================#
# Network plot            #
#=========================#
netgraph(mod)

}
