% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{netplot}
\alias{netplot}
\title{Produce a network plot}
\usage{
netplot(model, ...)
}
\arguments{
\item{model}{A \code{crossnmaModel} object produced by \code{crossnma.model}.}

\item{...}{\dots Additional arguments from \code{\link{netgraph.netmeta}}}
}
\description{
This function creates a plot of evidence network using an object produced by \code{crossnma.model}.
}
\examples{
# Two datasets
data(prt.data) #  participant-level data
data(std.data) # study-level data
 #=========================#
  # Create a jags model  #
 #=========================#
 # We conduct a network meta-analysis assuming a random-effects model.
 # The data comes from randomized-controlled trials and non-randomized studies (combined naively)
mod <- crossnma.model(prt.data=prt.data,
                      std.data=std.data,
                      trt="trt",
                      study="study",
                      outcome="outcome",
                      n="n",
                      design="design",
                      reference="A",
                      trt.effect="random",
                      covariate = NULL,
                      method.bias="naive"
                      )
 #=========================#
    # Network plot  #
 #=========================#
netplot(mod)

}
