% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossnma.plot.R
\name{plot.crossnma}
\alias{plot.crossnma}
\title{Generic plot function for crossnma.model object.}
\usage{
\method{plot}{crossnma}(x, ...)
}
\arguments{
\item{x}{The object generated by the crossnma.run function.}

\item{...}{\dots}
}
\description{
Generic plot function for crossnma.model object.
}
