% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.crossnma.model.R
\name{print.summary.crossnma.model}
\alias{print.summary.crossnma.model}
\title{Print code of JAGS model for cross-design & -format
network meta-analysis or regression}
\usage{
\method{print}{summary.crossnma.model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.crossnma.model}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
No return value (print function).
}
\description{
Print code of JAGS model for cross-design & -format
network meta-analysis or regression
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
stddata # study-level data

# Create a JAGS model
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,
  reference = "A", trt.effect = "random", method.bias = "naive")

summary(mod)

}
\seealso{
\code{\link{summary.crossnma.model}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{print}
