% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.crossnma.R
\name{print.summary.crossnma}
\alias{print.summary.crossnma}
\title{Print summary results for crossnma.run object}
\usage{
\method{print}{summary.crossnma}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{crossnma}.}

\item{digits}{The number of significant digits printed. The default
value is 3.}

\item{\dots}{Additional arguments.}
}
\description{
Print method for objects of class \code{crossnma}.
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

#=========================#
# Create a jags model     #
#=========================#
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,
  reference = "A", trt.effect = "random", method.bias = "naive")

#=========================#
# Fit jags model          #
#=========================#
fit <-
  crossnma.run(model = mod, n.adapt = 20,
    n.iter = 50, thin = 1, n.chains = 3)

#=========================#
# Display the output      #
#=========================#
print(summary(fit), digits = 5)

}
\seealso{
\code{\link{summary.crossnma}}
}
\author{
Tasnim Hamza \email{tasnim.hamza@ispm.unibe.ch}, Guido
  Schwarzer \email{sc@imbi.uni-freiburg.de}
}
\keyword{print}
