% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossnma.model.R
\name{crossnma.model}
\alias{crossnma.model}
\title{Create JAGS model to synthesize cross-design evidence and cross-format data in NMA and NMR for dichotomous outcomes}
\usage{
crossnma.model(
  prt.data,
  std.data,
  trt,
  study,
  outcome,
  n,
  design,
  reference = NULL,
  trt.effect = "random",
  covariate = NULL,
  reg0.effect = "independent",
  regb.effect = "random",
  regw.effect = "random",
  split.regcoef = T,
  method.bias = NULL,
  bias = NULL,
  bias.group = NULL,
  unfav = NULL,
  bias.type = NULL,
  bias.covariate = NULL,
  bias.effect = "common",
  down.wgt = NULL,
  prior = list(tau.trt = NULL, tau.reg0 = NULL, tau.regb = NULL, tau.regw = NULL,
    tau.gamma = NULL, pi.high.rct = NULL, pi.low.rct = NULL, pi.high.nrs = NULL,
    pi.low.nrs = NULL),
  run.nrs = list(var.infl = 1, mean.shift = 0, trt.effect = "common", n.adapt = 2000,
    n.iter = 10000, n.burnin = 4000, thin = 1, n.chains = 2)
)
}
\arguments{
\item{prt.data}{An object of class data.frame containing the individual participant dataset. Each row contains the data of a single participant.
The data frame needs to have the following columns: treatment, study identification, outcome (event and non-event), design. Additional columns might be required for certain analyses.}

\item{std.data}{An object of class data.frame containing the study-level dataset. Each row represents the information of study arm.
The data frame needs to have the following columns: treatment, study identification, outcome (number of events), sample size and design. Additional columns might be required for certain analyses.}

\item{trt}{A charachter of the name of the treatment variable in prt.data and std.data.}

\item{study}{A charachter of the name of the study variable in prt.data and std.data.}

\item{outcome}{A charachter of the name of the outcome variable in prt.data and std.data.}

\item{n}{A character of the name of the number of participants variable in std.data.}

\item{design}{A charachter of the name of the design variable in prt.data and std.data.}

\item{reference}{A character indicating the name of the reference treatment. When the reference is not specified, the first alphabetic treatment will be used as a reference in the analysis.}

\item{trt.effect}{A character defining the model for the study-specific treatment effects. Options are 'random' (default) or 'common'.}

\item{covariate}{An optional vector indicating the name of the covariates in prt.data and std.data (to conduct a network meta-regression)
The covariates can be either numeric or dichotomous variables. The user can provide up to 3 covariates.
For example, we set `covariate=c(‘age’, ‘sex’)` to adjust for 2 covariates.
The default option is `covariate=NULL` where no covariate adjustment is applied (network meta-analysis).}

\item{reg0.effect}{An optional character (when \code{covariate} is not NULL) indicating the relationship across studies for the prognostic effects expressed by the regression coefficient, (\eqn{\beta_0}), in a study \eqn{j}.
Options are 'independent' or 'random'. We recommend using 'independent' (default).}

\item{regb.effect}{An optional character (when \code{covariate} is not NULL) indicating the relationship across studies for the between-study regression coefficient (\eqn{\beta^B}). This parameter quantifies the treatment-mean covariate interaction.
Options are 'random' or 'common'. Default is 'random'.}

\item{regw.effect}{An optional character (when \code{covariate} is not NULL)  indicating the relationship across studies for the within-study regression coefficient (\eqn{\beta^W}). This parameter quantifies the treatment-covariate interaction effect at the individual level.
Options are 'random' and 'common'. Default is 'random'.}

\item{split.regcoef}{A logical value (when \code{covariate} is not NULL). If TRUE the within- and between-study coefficients will be splitted in the analysis of prt.data.
The default is TRUE. When the split.regcoef = FALSE, only a single regression coefficient will be estimated to represent both the between-studies and within-studies covariate effects.}

\item{method.bias}{A character for defining the method to combine randomised clinical trials (RCT) and non-randomised studies (NRS) (when \code{design} has both designs nrs and rct).
Options are 'naive' for naive synthesize, 'prior' for using NRS to inform priors for the relative treatment effects in RCTs.
or 'adjust1' and 'adjust2' to allow a bias adjustment. When only one design is available (either rct or nrs), this argument needs also to be specified to indicate whether unadjusted (naive) or bias-adjusted analysis (adjust1 or adjust2) should be applied.}

\item{bias}{A charachter indicating the name of the variable (required when method.bias='adjust1' or 'adjust2') that includes the risk of bias adjustment in prt.data and std.data.
The values of this variable should be a character with entries that need to be spelled as such 'low', 'high' or 'unclear'. These values need to be repeated for the participants that belong to the same study.}

\item{bias.group}{An optional charachter which defines the names of the variables in prt.data and std.data that indicates the bias effect in each study (can be provided when method.bias='adjust1' or 'adjust2')
The entries of these variables should be either 1 (study has inactive treatment and its estimate should be adjusted for bias effect), 2 (study has only active treatments and its estimate should be adjusted for bias effect (different from inactive bias effect)
or 0 (study doesn't need bias adjustment). The values need to be repeated for the participants that belong to the same study. Default is 1 which means applying bias adjustment to all study-specific estimates if the study is at high risk of bias.}

\item{unfav}{A charachter which defines the names of the variables in prt.data and std.data that include an indicator of the unfavoured treatment in each study (should be provided when method.bias='adjust1' or 'adjust2')
The entries of these variables should be either 0 (unfavoured treatment) or 1 (favourable treatment or treatments). Each study should include only one 0. The values need to be repeated for the participants that belong to the same study.}

\item{bias.type}{An optional character defining of bias on the treatment effect (required when method.bias='adjust1').
Three options are possible: 'add' to add the additive bias effect,'mult' for multiplicative bias effect and 'both' includes both an additive and a multiplicative terms.}

\item{bias.covariate}{A charachter of the variable name that will be used in estimating the probability of bias (can be provided when method.bias='adjust1' or 'adjust2')}

\item{bias.effect}{An optional character indicating the relationship for the bias coefficients across studies.
Options are 'random' or 'common' (default). It is required when method.bias='adjust1' or 'adjust2'.}

\item{down.wgt}{An optional numeric indicating the percent to which studies at high risk of bias will be down-weighted on average. The value ranges between 0 and 1. It can be provided when method.bias='adjust1' or 'adjust2'.}

\item{prior}{An optional list to control the prior for various parameters in JAGS model. When effects are set as 'random', we can set the heterogeneity parameters for: tau.trt for the treatment effects,
tau.reg0 for the effect of prognostic covariates, tau.regb and tau.regw for within- and between-study covariate effect, respectively.
and tau.gamma for bias effect. The default of all heterogeneity parameters is 'dunif(0,2)'. Currently only the uniform distribution is supported.
When the method.bias='adjust1' or 'adjust2', the user may provide priors to control the bias probability.
For the bias probabilities, beta distributions are assumed with the following default values: RCT with low (pi.low.rct='dbeta(1,10)')/high (pi.high.rct='dbeta(10,1)') bias, NRS with low(pi.low.rct='dbeta(1,30)')/high (pi.high.rct='dbeta(30,1)') bias (pi.low.nrs, pi.high.nrs).}

\item{run.nrs}{An optional list is needed when the NRS used as a prior (method.bias='prior').
The list consists of the following: (\code{var.infl}) controls the common inflation of the varaince of NRS estimates (\eqn{w}) and its values range between 0 (NRS do not contribute at all and the prior is vague) and 1 (the NRS evidence is used at face value, default approach).
The parameter (\code{mean.shift}) is the bias shift (\eqn{\zeta}) to be added/subtracted from the estimated mean treatment effects (on the log-scale) from NRS network (0 is the default). Either (\code{var.infl}) or (\code{mean.shift}) should be provided but not both.
\code{trt.effect} A character indicates how to combine treatment effects across studies .Options are 'random' or 'common' (default).
Here you can also specify the arguments to control the MCMC chains with default value is in the parentheses: the number of adaptions n.adapt (500), number of iterations n.iter(10000), number of burn in n.burnin (4000),
number of thinning thin (1) and number of chains n.chains (2), see \code{\link{jags.model}} arguments from rjags package.}
}
\value{
\code{crossnma.model} returns an object of class \code{crossnmaModel} which is a list containing the following components:

\code{jagsmodel}  A long character string containing JAGS code that will be run in \code{\link{jags}}.

\code{data}  The data used in the JAGS code.

\code{trt.key}  A table of the treatments and its mapped integer number (used in JAGS model).

\code{study.key}  A table of the studies and its mapped integer number (used in JAGS model).

\code{trt.effect} A character defining the model for the study-specific treatment effects.

\code{method.bias}  A character for defining the method to combine randomised clinical trials (RCT) and non-randomised studies (NRS).

\code{covariate}  A list of the the names of the covariates in prt.data and std.data used in network meta-regression.

\code{split.regcoef} A logical value. If FALSE the within- and between-study regression coefficients will be considered equal.

\code{regb.effect} An optional character  indicating the model for the between-study regression coefficients across studies.

\code{regw.effect} An optional character indicating the model for the within-study regression coefficients across studies.

\code{bias.effect} An optional character indicating the model for the bias coefficients across studies.

\code{bias.type} A character indicating the effect of bias on the treatment effect; additive ('add') or multiplicative ('mult') or both ('both').
}
\description{
This function creates a JAGS model and the needed data. The JAGS code is created from the internal function \code{crossnma.code}.
}
\seealso{
\code{\link{crossnma.run}}, \code{\link{jags.model}}
}
