% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crossnma.R
\name{plot.crossnma}
\alias{plot.crossnma}
\title{Trace plot of MCMC outputs.}
\usage{
\method{plot}{crossnma}(x, ...)
}
\arguments{
\item{x}{The object generated by the crossnma.run function.}

\item{...}{\dots}
}
\description{
Produces a separate plot for each parameter in the model. Each plot shows iterations vs sampled values. It takes crossnma.run object.
}
