% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{std.data}
\alias{std.data}
\title{A simulated aggregate data.
The dataset includes 2 randomised-controlled trials (RCTs), comparing 2 treatments: A, C.
The outcome is binary that is represented as the number of participants with a respond.}
\format{
An object of class \code{data.frame} with 4 rows and 11 columns.
}
\usage{
std.data
}
\description{

}
\details{

}
\keyword{datasets}
