% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.crossnma.model.R
\name{summary.crossnma.model}
\alias{summary.crossnma.model}
\title{Summary function for crossnma.model object}
\usage{
\method{summary}{crossnma.model}(x, ...)
}
\arguments{
\item{x}{An object generated by the
\code{\link{crossnma.model}}.}

\item{\dots}{Additional arguments (ignored)}
}
\description{
Summary function for crossnma.model object
}
\examples{
# We conduct a network meta-analysis assuming a random-effects
# model.
# The data comes from randomized-controlled trials and
# non-randomized studies (combined naively)
head(ipddata) # participant-level data
head(stddata) # study-level data

# Create a JAGS model
mod <- crossnma.model(treat, id, relapse, n, design,
  prt.data = ipddata, std.data = stddata,sm="OR",
  reference = "A", trt.effect = "random", method.bias = "naive")

summary(mod)

}
\seealso{
\code{\link{print.summary.crossnma.model}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
